Get email login working, start graphql api

+++ src/DotnetTests/Fixtures/FilledApplicationDbContextFixture.cs
@@ -2,6 +2,7 @@ using Microsoft.EntityFrameworkCore;
 using PersistenceService.Data.ApplicationDb;
 using PersistenceService.Data.SeedData;
 using DotnetTests.PersistenceService.Stores;
+using DotnetTest.SetupUtils;
 
 namespace DotnetTests.Fixtures;
 
@@ -11,52 +12,51 @@ public class FilledApplicationDbContextFixture : IAsyncLifetime
 
     private readonly TestSeeder _testSeeder;
 
+    private readonly bool _preserveSeededData;
+
+    private readonly bool _envIsDev;
+
     public FilledApplicationDbContextFixture()
     {
-        var envFilePath = Directory.GetCurrentDirectory() + "/../../../.env";
-        if (System.IO.File.Exists(envFilePath))
-        {
-            using (StreamReader reader = new StreamReader(envFilePath))
-            {
-                string? line;
-                while ((line = reader.ReadLine()) != null)
-                {
-                    int i = line.IndexOf("=");
-                    if (i == -1)
-                    {
-                        throw new InvalidOperationException(
-                            "Invalid .env file format"
-                        );
-                    }
-                    Environment.SetEnvironmentVariable(
-                        line.Substring(0, i),
-                        line.Substring(i + 1)
-                    );
-                }
-            }
-        }
+        SetupUtils.LoadEnvironmentVariables("/../../../.env");
+
+        _preserveSeededData =
+            Environment.GetEnvironmentVariable("PRESERVE_SEEDED_DATA")
+            == "true";
+        _envIsDev = Environment.GetEnvironmentVariable("ENV") == "dev";
+        string connectionString = _envIsDev
+            ? "LOCAL_DB_CONNECTION_STRING"
+            : "TEST_DB_CONNECTION_STRING";
         var options = new DbContextOptionsBuilder<ApplicationDbContext>()
-            .UseNpgsql(
-                Environment.GetEnvironmentVariable("TEST_DB_CONNECTION_STRING")
-            )
+            .UseNpgsql(Environment.GetEnvironmentVariable(connectionString))
             //.EnableSensitiveDataLogging()
             //.LogTo(Console.WriteLine)
             .Options;
 
         ApplicationDbContext c = new ApplicationDbContext(options);
-        c.Database.Migrate();
+        if (!c.Database.CanConnect())
+        {
+            c.Database.Migrate();
+        }
         context = c;
         _testSeeder = new TestSeeder(context, UserStoreTests.GetUserStore());
     }
 
     public async Task InitializeAsync()
     {
-        await _testSeeder.Seed(TestSeeder.Small);
+        string seedSize = Environment.GetEnvironmentVariable("LARGE_SEED_SIZE")
+            is null
+            ? TestSeeder.Small
+            : TestSeeder.Large;
+        await _testSeeder.Seed(seedSize);
     }
 
     public async Task DisposeAsync()
     {
-        context.Database.EnsureDeleted();
+        if (!_preserveSeededData)
+        {
+            context.Database.EnsureDeleted();
+        }
         await context.DisposeAsync();
     }
 }
